Fix compilation

+++ listeners/src/Main.sol
@@ -6,8 +6,8 @@ import "sim-idx-generated/Generated.sol";
 import "./BlockListener.sol";
 
 contract Triggers is BaseTriggers {
-    uint256 public constant BASE_START_BLOCK = 22576928;
-    uint256 public constant ETHEREUM_START_BLOCK = 21574092;
+    uint64 public constant BASE_START_BLOCK = 22576928;
+    uint64 public constant ETHEREUM_START_BLOCK = 21574092;
 
 
     function triggers() external virtual override {
